+++ engine/src/render/shape/render.rs
@@ -60,7 +60,7 @@ impl ShapeRenderPipeline {
 
         render_pass.set_bind_group(0, &self.bind_group, &[]);
         render_pass.set_vertex_buffer(0, self.vertex.slice(..));
-        render_pass.set_index_buffer(self.index.slice(..), IndexFormat::Uint16);
+        render_pass.set_index_buffer(self.index.slice(..), IndexFormat::Uint32);
         render_pass.draw_indexed(0..self.count, 0, 0..1);
     }
 }
